cloudstorage: access: added a wrapper for libcloudstorage functions calls
One of the invokations might be ugly because I had to specify which overloaded function I wanted.
There are two functions with the same name, which did not allow me to point to one of them without
disambiguate it. Either way, I will try to clean it up to make it readable