http: the connection will now sent a payload if requested
This was a necessary step in order to support POST, PUT, etc. The current code
had already support for custom headers and parameters, but not a payload on
the request message.

The payload field (stream) from the message is (was) always NULL when sent on
a request message, so it can be easily use the same way it is for reading
responses. I completed both versions (http1 and http2) in order to read that
payload and send it to the respective connection.